Ajout du service WeatherAPI.com qui n'a rien à voir avec Weather mais qui dérive de WWO. C'est un service assez intéressant qu'il convient d'essayer car il propose un plan gratuit correct et surtout les données WWO qu'il utilise a priori sont de bonnes qualité. Up de y
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@122899 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ rainette/trunk/inc/rainette_normaliser.php
@@ -62,7 +62,7 @@ $GLOBALS['rainette_config']['conditions'] = array(
 	'tendance_pression'     => array('origine' => 'service', 'type_php' => 'string', 'type_unite' => '', 'groupe' => 'donnees_atmospheriques'),
 	'visibilite'            => array('origine' => 'service', 'type_php' => 'float', 'type_unite' => 'distance', 'groupe' => 'donnees_atmospheriques'),
 	'nebulosite'            => array('origine' => 'service', 'type_php' => 'float', 'type_unite' => 'pourcentage', 'groupe' => 'donnees_atmospheriques'),
-	'indice_uv'             => array('origine' => 'service', 'type_php' => 'int', 'type_unite' => 'indice', 'groupe' => 'donnees_atmospheriques'),
+	'indice_uv'             => array('origine' => 'service', 'type_php' => 'float', 'type_unite' => 'indice', 'groupe' => 'donnees_atmospheriques'),
 	'risque_uv'             => array('origine' => 'calcul', 'type_php' => 'string', 'type_unite' => '', 'groupe' => 'donnees_atmospheriques'),
 	// Etats météorologiques natifs
 	'code_meteo'            => array('origine' => 'service', 'type_php' => 'string', 'type_unite' => '', 'groupe' => 'donnees_etats_natifs'),
@@ -99,7 +99,7 @@ $GLOBALS['rainette_config']['previsions'] = array(
 	'pression'             => array('origine' => 'service', 'type_php' => 'float', 'type_unite' => 'pression', 'rangement' => 'heure', 'groupe' => 'donnees_atmospheriques'),
 	'visibilite'           => array('origine' => 'service', 'type_php' => 'float', 'type_unite' => 'distance', 'rangement' => 'heure', 'groupe' => 'donnees_atmospheriques'),
 	'nebulosite'           => array('origine' => 'service', 'type_php' => 'float', 'type_unite' => 'pourcentage', 'rangement' => 'heure', 'groupe' => 'donnees_atmospheriques'),
-	'indice_uv'            => array('origine' => 'service', 'type_php' => 'int', 'type_unite' => 'indice', 'rangement' => 'heure', 'groupe' => 'donnees_atmospheriques'),
+	'indice_uv'            => array('origine' => 'service', 'type_php' => 'float', 'type_unite' => 'indice', 'rangement' => 'heure', 'groupe' => 'donnees_atmospheriques'),
 	'risque_uv'            => array('origine' => 'calcul', 'type_php' => 'string', 'type_unite' => '', 'rangement' => 'heure', 'groupe' => 'donnees_atmospheriques'),
 	// Etats météorologiques natifs
 	'code_meteo'           => array('origine' => 'service', 'type_php' => 'string', 'type_unite' => '', 'rangement' => 'heure', 'groupe' => 'donnees_etats_natifs'),
